+++ Structurizr.Core/View/ComponentView.cs
@@ -42,6 +42,12 @@ namespace Structurizr
                 this.containerId = value;
             }
         }
+        
+        /// <summary>
+        /// Determines whether container boundaries should be visible for "external" components (those outside the container in scope).
+        /// </summary>
+        [DataMember(Name = "externalContainerBoundariesVisible", EmitDefaultValue = false)]
+        public bool? ExternalContainerBoundariesVisible { get; set; }
 
         internal ComponentView() : base()
         {
@@ -101,10 +107,7 @@ namespace Structurizr
         {
             if (component != null)
             {
-                if (Container.Equals(component.Container))
-                {
-                    AddElement(component, true);
-                }
+                AddElement(component, true);
             }
         }