+++ Structurizr.Core/View/Routing.cs
@@ -3,6 +3,7 @@
     public enum Routing
     {
         Direct,
+        Curved,
         Orthogonal
     }
 }