+++ docs/changelog.md
@@ -3,6 +3,16 @@
 ## 0.9.7 (unreleased)
 
 - Added an ExternalSoftwareSystemBoundariesVisible property to ContainerView, to set whether software system boundaries should be visible for "external" containers (those outside the software system in scope).
+- Added an ExternalContainersBoundariesVisible property to ComponentView, to set whether container boundaries should be visible for "external" components (those outside the container in scope).
+- Added a 16:10 ratio paper size.
+- Added a "Component" element shape.
+- Added a "Dotted" element border style.
+- Components from any container can now be added to a component view.
+- Added the ability to customize the symbols used when rendering metadata.
+- Adds support for curved relationship routing.
+- Adds support for multiple themes.
+- Adds support for infrastructure nodes.
+- Adds support for animation steps on deployment views.
 
 ## 0.9.6 (29th February 2020)